add init function for home and categories pages
