+++ src/style.css
@@ -290,7 +290,7 @@ header nav ul {
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.4); /* transparent black */
-  z-index: 1;
+  z-index: 0;
 }
 .hero-text {
   z-index: 10;
@@ -679,7 +679,7 @@ footer p {
   position: absolute;
   top: 15%;
   right: 2%;
-  z-index: 9999;
+  z-index: 12;
   width: 32rem;
   background-color: var(--color-white);
   color: var(--color-black);
